+++ cpp/include/cudf/io_types.h
@@ -77,14 +77,19 @@ typedef struct {
   gdf_size_type skiprows;                   ///< Number of rows at the start of the files to skip, default is 0
   gdf_size_type skipfooter;                 ///< Number of rows at the bottom of the file to skip - default is 0
 
-  bool          skip_blank_lines;           // Whether or not to ignore blank lines
+  bool          skip_blank_lines;           ///< Indicates whether to ignore empty lines, or parse and interpret values as NaN 
 
   const char    **true_values;              ///< List of values to recognize as boolean True
   int           num_true_values;            ///< Number of values in the true_values list
   const char    **false_values;             ///< List of values to recognize as boolean False
   int           num_false_values;           ///< Number of values in the true_values list
 
-  const char    **na_values;                // Array of strings that should be considered as NA
+  const char    **na_values;                /**< Array of strings that should be considered as NA. By default the following values are interpreted as NaN: 
+                                            '', '#N/A', '#N/A N/A', '#NA', '-1.#IND', '-1.#QNAN', '-NaN', '-nan', '1.#IND', '1.#QNAN', 'N/A', 'NA', 'NULL',
+                                            'NaN', 'n/a', 'nan', 'null'. */
+  int           num_na_values;              ///< Number of values in the na_values list
+  bool          keep_default_na;            ///< Keep the default NA values
+  bool          na_filter;                  ///< Detect missing values (empty strings and the values in na_values). Passing false can improve performance.
 
   char          *prefix;                    // If there is no header or names, append this to the column ID as the name
   bool          mangle_dupe_cols;           ///< If true, duplicate columns get a suffix. If false, data will be overwritten if there are columns with duplicate names
@@ -104,7 +109,7 @@ typedef struct {
 
   char          escapechar;                 // Single character used as the escape character
 
-  char          comment;                    // Single character indicating that the remainder of line is a comment
+  char          comment;                    ///< The character used to denote start of a comment line. The rest of the line will not be parsed.
 
   char          *encoding;                  // the data encoding, NULL = UTF-8
 
@@ -120,8 +125,6 @@ typedef struct {
  *
  * squeeze          - data is always returned as a gdf_column array
  * engine           - this is the only engine
- * keep_default_na  - this has no meaning since the field is marked invalid and the value not seen
- * na_filter        - empty fields are automatically tagged as invalid
  * verbose
  * keep_date_col    - will not maintain raw data
  * date_parser      - there is only this parser